usb: typec: tcpm: add vdm retry mechanism
Vendor Defined Message (Discover Identity,Discover SVIDs,Discover Modes,so on)
is queued to VDM state machine immediately in Ready states and never retries
if it fails in the AMS. If VDM is blocked by Collision Avoidance mechanism,
Sink needs to wait and retry again when CC is idle. For example, Sink can start
AMS only when SinkTxOk is set on the Source side.

As shown in the log below, sink starts VDM in SinkTxNG state, and the DP detection
is abnormally closed after the failure.

[   35.687167] state change DR_SWAP_ACCEPT -> DR_SWAP_CHANGE_DR [rev3 NONE_AMS]
[   35.687188] Requesting mux state 1, usb-role 1, orientation 2
[   35.688471] state change DR_SWAP_CHANGE_DR -> SNK_READY [rev3 NONE_AMS]
[   35.689008] CC1: 0 -> 0, CC2: 4 -> 5 [state SNK_READY, polarity 1, connected] <--> [Source set SinkTxOk]
[   35.689027] AMS DISCOVER_IDENTITY start <--> [Sink start DISCOVER_IDENTITY in SinkTxOk, result: success]
[   35.689035] PD TX, header: 0x16af
[   35.691368] PD TX complete, status: 0
[   35.693665] CC1: 0 -> 0, CC2: 5 -> 4 [state SNK_READY, polarity 1, connected] <--> [Source set SinkTxNG]
[   35.694947] PD RX, header: 0x5b8f [1]
[   35.694958] Rx VDM cmd 0xff00a041 type 1 cmd 1 len 5
[   35.694966] AMS DISCOVER_IDENTITY finished
[   35.695019] Identity: 2109:0102.0001
[   35.695033] AMS DISCOVER_SVIDS start    <--> [Sink start DISCOVER_SVIDS in SinkTxNG, result: failed]
[   35.695038] Sink TX No Go
[   35.695876] CC1: 0 -> 0, CC2: 4 -> 5 [state SNK_READY, polarity 1, connected] <--> [Source set SinkTxOk]

Change-Id: Id5b0f91edb11fd10187f3ac49d03c5e53f8130c2
Signed-off-by: Wang Jie <dave.wang@rock-chips.com>